% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{add_default_rec}
\alias{add_default_rec}
\title{Add a default data line for each point source for which no user input data
was provided in the data folder.}
\usage{
add_default_rec(tbl, ob_names, rec_col_names)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}

\item{ob_names}{Name vector of point sources for which a default point source
data entry should be generated.}

\item{rec_col_names}{Character vector with recall data column names}
}
\description{
The default data is a constant input with all zeros for all variables.
}
\keyword{internal}
