% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_latlon}
\alias{add_latlon}
\title{Add lat lon coordinates in WGS84 to the vector file attribute table.}
\usage{
add_latlon(layer)
}
\arguments{
\item{layer}{Vector layer}
}
\description{
Add lat lon coordinates in WGS84 to the vector file attribute table.
}
\keyword{internal}
