% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{add_missing_cols}
\alias{add_missing_cols}
\title{Add the columns which are part of the recall data table but were not provided
as input to the user provided point source data.}
\usage{
add_missing_cols(tbl, rec_col_names)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}

\item{rec_col_names}{Character vector with recall data column names}
}
\description{
Add the columns which are part of the recall data table but were not provided
as input to the user provided point source data.
}
\keyword{internal}
