% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{add_point_sources}
\alias{add_point_sources}
\title{Add point sources (recall) to the SWAT+ model setup.}
\usage{
add_point_sources(point_path, data_path, max_point_dist)
}
\arguments{
\item{point_path}{Path of the point source point shape layer}

\item{data_path}{Path of the project data folder}

\item{max_point_dist}{Maximum distance to snap a point source to the closest
channel or reservoir object. Numeric value in meters.}
}
\value{
Writes the following tables into tables.sqlite:
\itemize{
\item recall.recall_con
\item recall.recall_con_out
\item recall.recall_dat
\item recall.recall_rec
}
}
\description{
The point source locations are derived from a point shape file input layer.
Optionally point source records are added if csv files with the same name as
the name of a point source exist in the same folder with the point source
location layer
}
