% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetland.R
\name{add_wetlands}
\alias{add_wetlands}
\title{Build the wetland input tables and add surface water storage to HRUs.}
\usage{
add_wetlands(data_path, wetland_landuse)
}
\arguments{
\item{data_path}{Path of the project data folder}

\item{wetland_landuse}{Character vector that defines the wetland land uses
for which surface water storage is initialized.}
}
\value{
The function writes the following tables into tables.sqlite:
\itemize{
\item wet.wetland_wet
\item wet.hydrology_wet
}

The function additionally rewrites hru.hru_data_hru and adds surf_stor_ids
in case wetland land uses are used in the land layer.
}
\description{
Build the wetland input tables and add surface water storage to HRUs.
}
