% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_soil.R
\name{aggregate_hru_dem_soil}
\alias{aggregate_hru_dem_soil}
\title{Aggregate the raster properties of the DEM and the soil layer for the
units in the hru layer}
\usage{
aggregate_hru_dem_soil(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
Writes the tabble hru_terrain_soil into the data base tables.sqlite.
}
\description{
Aggregate the raster properties of the DEM and the soil layer for the
units in the hru layer
}
