% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_na}
\alias{all_na}
\title{Identifies if all elements of a vector are NA}
\usage{
all_na(x)
}
\arguments{
\item{x}{Vector of any type}
}
\description{
Identifies if all elements of a vector are NA
}
\keyword{internal}
