% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hru.R
\name{assign_potential}
\alias{assign_potential}
\title{Assign the leaching or the runoff potential to an HRU based on the
provided lookup table, the slope and the HSG.}
\usage{
assign_potential(slp, hsg, lookup)
}
\arguments{
\item{slp}{Numeric value, slope in m per m}

\item{hsg}{Character value, one of 'A', 'B', 'C', or 'D'}

\item{lookup}{Tibble that gives the translation to the potential based
slope and HSG}
}
\description{
Assign the leaching or the runoff potential to an HRU based on the
provided lookup table, the slope and the HSG.
}
\keyword{internal}
