% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer.R
\name{build_cha_id_list}
\alias{build_cha_id_list}
\title{Build the sequence of elements (channel ids) which is required in the
aqu_cha.lin file to connect the aquifer and the channels.}
\usage{
build_cha_id_list(cha_id)
}
\arguments{
\item{cha_id}{Numeric vector with channel ids}
}
\description{
Build the sequence of elements (channel ids) which is required in the
aqu_cha.lin file to connect the aquifer and the channels.
}
\keyword{internal}
