% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_res.R
\name{build_cha_input}
\alias{build_cha_input}
\title{Build SWAT+ channel input files.}
\usage{
build_cha_input(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The function writes the following tables into the intermediate
data base:
\itemize{
\item cha.chandeg_con
\item cha.hyd_sed_lte_cha
\item cha.channel_lte_cha
}
}
\description{
Build SWAT+ channel input files.
}
