% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hru.R
\name{build_landuse}
\alias{build_landuse}
\title{Build the landuse_lum and the hru land use id table.}
\usage{
build_landuse(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The function writes the following tables into the intermediate
data base:
\itemize{
\item landuse.ids
\item landuse.landuse_lum
}
}
\description{
Build the landuse_lum and the hru land use id table.
}
