% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtu_lsu.R
\name{build_ls_unit_input}
\alias{build_ls_unit_input}
\title{Build SWAT+ landscape unit inut files.}
\usage{
build_ls_unit_input(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The function writes the following tables into the intermediate data
base:
\itemize{
\item lsu.ls_unit_def
\item lsu.ls_unit_ele
}
}
\description{
Build SWAT+ landscape unit inut files.
}
