% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_res.R
\name{build_res_input}
\alias{build_res_input}
\title{Build SWAT+ reservoir input files.}
\usage{
build_res_input(data_path)
}
\arguments{
\item{data_path}{Path to the project data folder}
}
\value{
The function writes the following tables into the intermediate
data base:
\itemize{
\item res.reservoir_con
\item res.hydrology_res
\item res.reservoir_res
}
}
\description{
Build SWAT+ reservoir input files.
}
