% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtu_lsu.R
\name{build_rout_con_out}
\alias{build_rout_con_out}
\title{Build SWAT+ routing unit con_out based on the connect_ids table.}
\usage{
build_rout_con_out(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\description{
This already adds the connection to the single aquifer that will be added to
the setup in a next step.
}
