% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtu_lsu.R
\name{build_rout_input}
\alias{build_rout_input}
\title{Build SWAT+ routing unit input files.}
\usage{
build_rout_input(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The function writes the following tables into the intermediate data
base:
\itemize{
\item rtu.rout_unit_con
\item rtu.rout_unit_ele
\item rtu.rout_unit_rtu
\item rtu.field_fld
}
}
\description{
In the current configuration of SWATbuildR setups one routing unit only
contains one HRU. Hence, the routing unit inputs are generated from the hru
input tables.
}
