% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer.R
\name{build_single_aquifer_files}
\alias{build_single_aquifer_files}
\title{Build the SWAT+ aquifer input files for a single aquifer for the
entire catchment.}
\usage{
build_single_aquifer_files(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The function writes the following tables into the intermediate
data base:
\itemize{
\item aqu.aquifer_con
\item aqu.aquifer_aqu
}
}
\description{
Build the SWAT+ aquifer input files for a single aquifer for the
entire catchment.
}
