% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_soil.R
\name{build_soil_data}
\alias{build_soil_data}
\title{Read and prepare the soil data.}
\usage{
build_soil_data(soil_lookup_path, soil_data_path, data_path)
}
\arguments{
\item{soil_lookup_path}{Path of the soil lookup csv file}

\item{soil_data_path}{Path of the soil data csv file}

\item{data_path}{Path of the project data folder}
}
\value{
The function writes 4 tables into tables.sqlite:
\itemize{
\item soil.lookup          provides the soil lookup table
\item soil.soils_sol       provides the soils_sol table
\item soil.soils_sol_layer provides the soils_sol layer information
\item .$ids                provides the ID links between HRU ids and soil ids
}
}
\description{
Read and prepare the soil data.
}
