% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_connectivity.R
\name{build_water_object_connectivity}
\alias{build_water_object_connectivity}
\title{Calculate the water object connectivity.}
\usage{
build_water_object_connectivity(data_path)
}
\arguments{
\item{vct_path}{Character string, path to the project vector files}

\item{id_cha_out}{Integer value, if set the respective channel unit is
considered as the end unit of the water object network}

\item{id_res_out}{Integer value, if set the respective reservoir unit is
considered as the end unit of the water object network}
}
\value{
Writes the following tables into the intermediate data base
data/tables.sqlite:
\itemize{
\item cha.chandeg_con_out
\item res.reservoir_con_out
}
}
\description{
The routine identifies the connections between channel and reservoir objects.
It propagates from the catchment outlet object backwards to link all objects
of the network. The modeler has to provide either the id of the final channel
or the final reservoir object (initial ids of the input layers channel or
land). The routine returns a list with the 'chandeg_con_out' and the
'reservoir_con_out' tables
}
