% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_characteristics.R
\name{calc_wd}
\alias{calc_wd}
\title{Calculate channel width, depth and width to depth ratio based on the
contributing area of a channel.}
\usage{
calc_wd(contr_area)
}
\arguments{
\item{contr_area}{Contributing area of a channel as a numeric value.}
}
\description{
Calculate channel width, depth and width to depth ratio based on the
contributing area of a channel.
}
\keyword{internal}
