% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_characteristics.R
\name{calculate_contributing_area}
\alias{calculate_contributing_area}
\title{Calculate the contributing area of channels. This function is a more
robust replacement for a simple flow accumulation calculation.
This function calculates iso-basins with a size of 500m. The flow
accumulation for the iso-basins is calculated.
For each channel quantiles of flow accumultation values are calculated.
First and second order derivatives are calculated for the differences
between quantiles. If the change in accumulation between quantiles is
similar (low sd) it is assumed that flow accumulation steadily increases
for a channel. Then the largest value of flow accumulation is used that
is cut by the channel. If the difference between quantiles is low then
the median quantile value is used. If the quantile values provide a
rather indifferent picture the flow accumulation is selected which meets
a plausible value considering the flow order of that channel (Large order
higher accumulation value, low order tendency to lower accumulation value)}
\usage{
calculate_contributing_area(data_path, cha_rst, dem_res)
}
\arguments{
\item{data_path}{Path to the project data folder}

\item{cha_rst}{Channel id raster layer}

\item{dem_res}{Spatial resolution of the DEM. Single numeric value of (x,y)
cell size in meters.}
}
\description{
Calculate the contributing area of channels. This function is a more
robust replacement for a simple flow accumulation calculation.
This function calculates iso-basins with a size of 500m. The flow
accumulation for the iso-basins is calculated.
For each channel quantiles of flow accumultation values are calculated.
First and second order derivatives are calculated for the differences
between quantiles. If the change in accumulation between quantiles is
similar (low sd) it is assumed that flow accumulation steadily increases
for a channel. Then the largest value of flow accumulation is used that
is cut by the channel. If the difference between quantiles is low then
the median quantile value is used. If the quantile values provide a
rather indifferent picture the flow accumulation is selected which meets
a plausible value considering the flow order of that channel (Large order
higher accumulation value, low order tendency to lower accumulation value)
}
\keyword{internal}
