% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_connectivity.R
\name{check_cha_res_connectivity}
\alias{check_cha_res_connectivity}
\title{Check the water object connectivity.}
\usage{
check_cha_res_connectivity(data_path, id_cha_out, id_res_out)
}
\arguments{
\item{data_path}{Character string, path to the project data folder}

\item{id_cha_out}{Integer value, if set the respective channel unit is
considered as the end unit of the water object network}

\item{id_res_out}{Integer value, if set the respective reservoir unit is
considered as the end unit of the water object network}

\item{cha}{Vector layer of the channels}
}
\value{
Writes geopackage files into data/vector with layers which show
potential issues in the object connectivity.
}
\description{
The routine identifies any issues with the connectivity between channel and
reservoir objects. The function prompts check messages in the command line.
If issues were identified respective channel and reservoir layers are written
into the vector data path in 'water_connectivity_issues.gpkg'
}
