% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite_loops.R
\name{check_infinite_loops}
\alias{check_infinite_loops}
\title{Identify infinite loops in the object connections.}
\usage{
check_infinite_loops(connect_ids, data_path, type, n_iter = 50)
}
\arguments{
\item{connect_ids}{Table with the from-to object connectivities and the
respective net flow accumulations.}

\item{data_path}{Path of the project data folder}

\item{type}{String indicating if 'Land' or 'Water' objects are analyzed.}

\item{n_iter}{Number of iterations = length of the connections
which are analyzed.}
}
\description{
Analyze the calculated connections between land objects or water
objects for infinite loop routing by walking through the connections
starting from each spatial object and analyzing if any connection
returns to the initial connection.
In the case of identified loop routing the routine writes the affected
spatial objects into a geopackage for further analysis.
}
