% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_land_drain_ids}
\alias{check_land_drain_ids}
\title{Check the IDs of draining land objects to channel objects.}
\usage{
check_land_drain_ids(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\description{
In the case of drainage (e.g. tile or total flow) from a land object to a
channel object, it is necessary to check if the defined channel ids exist.
This routine performs checks for the defined drainage ids in the land layer.
}
