% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_layer_attributes}
\alias{check_layer_attributes}
\title{Check attribute columns of read input layers.}
\usage{
check_layer_attributes(layer, type_to_lower)
}
\arguments{
\item{layer}{Vector polygon layer as sf object.}

\item{type_to_lower}{Boolean input to define if the type labels should be
converted to all lower case labels.}
}
\value{
The checked layer as sf object with renamed columns.
}
\description{
Check if the layer attributes at least contain the names 'id' and 'type'.
Further check if 'ids' are unique and convert to lower case 'type' if
wanted.
}
