% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_project_crs}
\alias{check_project_crs}
\title{Check if the reference system of the project and a layer differ}
\usage{
check_project_crs(
  layer,
  data_path,
  proj_layer,
  label,
  type,
  method = "bilinear"
)
}
\arguments{
\item{layer}{GIS layer for which the reference system is compared to the project
CRS}

\item{data_path}{Path of the project data folder}

\item{proj_layer}{Boolean input to define if the layer should be projected.}

\item{label}{Layer name which is  printed in case of an error message}

\item{type}{String that defines layer type 'raster' for terra raster layers
and 'shp' for sf vector layers}

\item{method}{Only used for raster. Method used for estimation of new cell
values when projecting. E.g. use 'near' for soil classes and 'bilinear'
(default) for elevation.}
}
\value{
The transformed input layer as sf object
}
\description{
Check if the reference system of the project and a layer differ
}
