% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_raster_coverage}
\alias{check_raster_coverage}
\title{Check the coverage of a raster layer input for the land layer polygons}
\usage{
check_raster_coverage(rst, vct_layer, data_path, label, cov_frc)
}
\arguments{
\item{rst}{Raster layer}

\item{vct_layer}{Name of the vector layer to analyze coverage for.}

\item{data_path}{Path of the project data folder}

\item{label}{Layer name to print in error message}

\item{cov_frc}{Threshold value of area fraction of coverage for each polygon
feature to be covered by the raster layer.}
}
\value{
The functions returns an error if the coverage of a polygon is < cov_frc.
}
\description{
Check the coverage of a raster layer input for the land layer polygons
}
