% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{check_unique_time}
\alias{check_unique_time}
\title{Check if all provided time steps are unique and no duplicate time steps are
provided in the user input point source table}
\usage{
check_unique_time(tbl, tbl_name)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}

\item{tbl_name}{Name of the analyzed point source}
}
\description{
Check if all provided time steps are unique and no duplicate time steps are
provided in the user input point source table
}
\keyword{internal}
