% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{check_var_names}
\alias{check_var_names}
\title{Check the variable names in the user provided point source data tables.
The function triggers an error message showing for all input tables the
column names which are incorrect point source variable names.}
\usage{
check_var_names(pnt_list, rec_col_names)
}
\arguments{
\item{pnt_list}{List of tibbles of user input point source data.}

\item{rec_col_names}{Character vector with recall data column names}
}
\description{
Check the variable names in the user provided point source data tables.
The function triggers an error message showing for all input tables the
column names which are incorrect point source variable names.
}
\keyword{internal}
