% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_vector_coverage}
\alias{check_vector_coverage}
\title{Check the area coverage of a polygon vector layer compared to the area
of the basin boundary.}
\usage{
check_vector_coverage(layer, bound, label, coverage)
}
\arguments{
\item{layer}{Polygon vector layer}

\item{bound}{Basin boundary layer}

\item{label}{Layer name to print in error message}

\item{coverage}{Numeric coverage threshold value in percent.}
}
\description{
The function triggers an error if the area fraction
of 'shp' to 'bound' is lower than the value defined with 'coverage'.
}
