% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{check_ymd_cols}
\alias{check_ymd_cols}
\title{Check if all required date columns are provided in a point source input table.}
\usage{
check_ymd_cols(tbl, tbl_name)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}

\item{tbl_name}{Name of the analyzed point source}
}
\description{
Check if all required date columns are provided in a point source input table.
}
\keyword{internal}
