% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite_loops.R
\name{count_pattern}
\alias{count_pattern}
\title{Count in how many of the 'sequences' the 'pattern' can be found.}
\usage{
count_pattern(pattern, sequences)
}
\arguments{
\item{pattern}{Text string with pattern of 'from_id,to_id'.}

\item{sequences}{Text string with id sequences in the connectivity}
}
\description{
Count in how many of the 'sequences' the 'pattern' can be found.
}
\keyword{internal}
