% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_names}
\alias{create_names}
\title{Create object names based on the 'id' column of the attribute table}
\usage{
create_names(id, prefix)
}
\arguments{
\item{id}{Numeric id vector of the attribute table}

\item{prefix}{character prefix e.g. 'hru', 'cha', etc.}
}
\description{
Create object names based on the 'id' column of the attribute table
}
\keyword{internal}
