% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_base.R
\name{create_swatplus_database}
\alias{create_swatplus_database}
\title{Write the SWAT+Editor project database.}
\usage{
create_swatplus_database(project_path, project_name)
}
\arguments{
\item{project_path}{Path of the project folder}

\item{project_name}{User defined name of the project}
}
\value{
An SQLite data base file with the name '<project_name>.sqlite' that
can be further edited with the SWAT+Editor.
}
\description{
The database will be located in the 'project_path' with the name
'<project_name>.sqlite'.
After writing the database it can be opened and edited with the SWAT+Editor.
}
