% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{date_to_ymd}
\alias{date_to_ymd}
\title{Conversion from date column to julian day, day, month, year format.}
\usage{
date_to_ymd(tbl)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}
}
\description{
Conversion from date column to julian day, day, month, year format.
}
\keyword{internal}
