% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_connectivity.R
\name{find_drain_id}
\alias{find_drain_id}
\title{Identify drain IDs}
\usage{
find_drain_id(id_mtx, bnd_mtx, acc_mtx, pnt_mtx, id_i)
}
\arguments{
\item{id_mtx}{id matrix created from the neighborhood raster of object i}

\item{bnd_mtx}{borderline matrix created from the border raster of object i}

\item{acc_mtx}{flow accumulation matrix for object i}

\item{pnt_mtx}{flow pointer matrix for object i}

\item{id_i}{Numeric value that gives the id value of object i}
}
\value{
A table with the from and to object IDs with the corresponding
flow accumulations
}
\description{
Identify the ids and calculate their flow_accumulation sums for the
objects to which a land object drains its water based on its flow
accumulation and its flow pointer.
}
\keyword{internal}
