% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_connectivity.R
\name{generate_con_out}
\alias{generate_con_out}
\title{Generate con_out tables from the connectivity link tables.}
\usage{
generate_con_out(con_tbl)
}
\arguments{
\item{con_tbl}{Connectivity link table}
}
\description{
The connectivity link tables are translated into the SWAT input table format
of *_con_out tables.
}
\keyword{internal}
