% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_line_endpoints}
\alias{get_line_endpoints}
\title{Extract start and end points of line features.}
\usage{
get_line_endpoints(layer)
}
\arguments{
\item{layer}{Vector linestring layer}
}
\description{
The function returns a list with two table start and end, holding the start
and end points of the line features as point features.
}
\keyword{internal}
