% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{get_min_dist}
\alias{get_min_dist}
\title{Find the minimum distance of a point to another vector object.}
\usage{
get_min_dist(pnt, obj, typ)
}
\arguments{
\item{pnt}{Point source vector layer as sf object}

\item{obj}{Object input layer as sf object for which the closest feature to a
point is identified.}

\item{typ}{Character label that defines the type of object.}
}
\value{
A tibble with the closest distances of a point source to a water
object.
}
\description{
The function returns a tibble with the point id and the object id and
distance to the nearest feature of a second vector input layer.
}
\keyword{internal}
