% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_connectivity.R
\name{ind2sub}
\alias{ind2sub}
\title{Conversion of index value to matrix row column coordinates}
\usage{
ind2sub(ind, n_r)
}
\arguments{
\item{ind}{Numeric value of vector index}

\item{n_r}{Numeric value, number of rows of the matrix}
}
\value{
vector of length 2 with row/column indices.
}
\description{
Conversion of index value to matrix row column coordinates
}
\keyword{internal}
