% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_obj_layer}
\alias{init_obj_layer}
\title{Initialize the base structure of the object vector layer attribute
tables.}
\usage{
init_obj_layer(layer, name_prefix)
}
\arguments{
\item{layer}{Vector layer}

\item{name_prefix}{Character prefix e.g. 'hru', 'cha', etc.}
}
\description{
Initialize the base structure of the object vector layer attribute
tables.
}
\keyword{internal}
