% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_buildr.R
\name{initialize_buildr}
\alias{initialize_buildr}
\title{Initialize a new SWATbuildR project}
\usage{
initialize_buildr(project_name, path, start_new_session = FALSE)
}
\arguments{
\item{project_name}{Name of the new SWATbuildR project (This will be the name
of the project folder).}

\item{path}{Path where the hard project folder should be initialized.}

\item{start_new_session}{Open R project in new or the current R session.}
}
\value{
Generates and opens a template for a new SWATbuildR project.
}
\description{
\code{initialize_buildr()} initializes a template SWATbuildR project. The routine
creates a new project folder with the name \code{project_name} in \code{path}. The
folder contains the complete structure of a SWATbuildR project and contains
demo data which must be replaced with your own input data. The input files
can be named differently than the demo data. The file paths of the renamed
files must be defined in the settings file \emph{settings.R}. The scripted
workflow to build a SWATbuildR model setup is then available in the R script
\emph{swatbuildr.R} which is located in the generated project folder.
}
\examples{
\dontrun{
initialize_buildr('test_project', '~/', start_new_session = TRUE)
}

}
