% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hru.R
\name{initialize_perco_cn3_latq}
\alias{initialize_perco_cn3_latq}
\title{Initialize the HRU parameters perco, cn3_swf, and latq_co based on the
leaching potential and the runoff potential of an HRU.}
\usage{
initialize_perco_cn3_latq(hru_terrain_soil, soil, is_drained)
}
\arguments{
\item{hru_terrain_soil}{Table that links terrain properties with hru ids}

\item{soil}{List of tables that provides the read soil data
(\code{soil$data}) and the links between soil ids with hru ids
(\code{soil$ids})}

\item{is_drained}{Boolean vector which indicates for each hru if it is
drained or not.}
}
\description{
The leaching potential and the runoff potential can be identified based on the
average slope of an HRU and the hydrological soil group (HSG) of an HRU.
The classification of the leaching and runoff potentials were done
according to Thompson et al. (2020).
}
\keyword{internal}
