% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer.R
\name{link_aquifer_channels}
\alias{link_aquifer_channels}
\title{Link the aquifer and the channels to simulate recharge from the aquifers
to the channels with geomorphic flow.}
\usage{
link_aquifer_channels(txt_path)
}
\arguments{
\item{txt_path}{Path of the projects TxtInOut folder}
}
\value{
The function adds the SWAT+ input file aqu_cha.lin which links
aquifer ids and channel ids and updates the file.cio to point to this
link file.
}
\description{
This function is used to update the SWAT+ text input files which were
written with the SWAT+Editor to activate the geomorphic flow option
to simulate recharge from the aquifer into the channels.
}
