% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite_loops.R
\name{match_pair}
\alias{match_pair}
\title{Identify if the pattern of a connection was already found before in this
connection chain.}
\usage{
match_pair(vct)
}
\arguments{
\item{vct}{Numeric vector of object ids in a connection chain.}
}
\description{
This is used to remove connection duplicates and reduce the number of
analyzed connections.
}
\keyword{internal}
