% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode}
\alias{mode}
\title{Calculate the modal value of a vector}
\usage{
mode(v, na.rm = TRUE)
}
\arguments{
\item{v}{Numeric vector}
}
\description{
Calculate the modal value of a vector
}
\keyword{internal}
