% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{prepare_recall_data}
\alias{prepare_recall_data}
\title{Prepare point source time series data.}
\usage{
prepare_recall_data(point_path, recall_con)
}
\arguments{
\item{point_path}{Path of the point source point shape layer}

\item{recall_con}{Generated recall_con tibble}
}
\value{
A list with the two tables recall_dat and recall_rec.
}
\description{
Prepare the time series input data from the point source data folder and
generate the recall data and the recall record type tables. The tables are
returned as list with elements recall_dat and recall_rec
}
\keyword{internal}
