% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_connectivity.R
\name{prepare_terrain_land}
\alias{prepare_terrain_land}
\title{Prepare terrain properties for the land objects.}
\usage{
prepare_terrain_land(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
Writes hydrologically corrected flow pointer and flow accumulation
raster layers into data/raster.
}
\description{
Prepare raster layers based on the DEM and the surface channel objects
that will be used in the calculation of the land object connectivity.
The routine writes the prepared raster layer into the raster folder
of the project.
}
