% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_res.R
\name{prepare_terrain_water}
\alias{prepare_terrain_water}
\title{Calculate terrain properties such as elevation, slope, catchment area,
channel width/depth for channel and reservoir objects}
\usage{
prepare_terrain_water(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
A list with the tibbles 'cha' and 'res' that provide the
respective terrain properties for the two spatial object types.
}
\description{
Calculate terrain properties such as elevation, slope, catchment area,
channel width/depth for channel and reservoir objects
}
