% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_connectivity.R
\name{prepare_water_links}
\alias{prepare_water_links}
\title{Prepare id_from and id_to links for the water objects to be used in
the infinite loops check for water objects.}
\usage{
prepare_water_links(data_path)
}
\arguments{
\item{data_path}{List of tables with cha and res con_out tables}
}
\description{
Prepare id_from and id_to links for the water objects to be used in
the infinite loops check for water objects.
}
