% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_connectivity.R
\name{prepare_water_objects}
\alias{prepare_water_objects}
\title{Prepare channel and reservoir objects for the generation of the water object
network.}
\usage{
prepare_water_objects(vct_path)
}
\arguments{
\item{data_path}{Character string, path to the project data folder}
}
\value{
A list with channel and reserwoir object tables.
}
\description{
The reservoir and the channel layer attribute tables are restructured to have
an id from 1 to nrow(.) and keeping the initial object ids as 'id_sel', also
object names and empty gis_id are added.
}
\keyword{internal}
