% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_plural}
\alias{print_plural}
\title{Print word with plural 's' if n is not 1}
\usage{
print_plural(word, n)
}
\arguments{
\item{word}{Text string}

\item{n}{Integer value}
}
\description{
Print word with plural 's' if n is not 1
}
\keyword{internal}
