% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_base.R
\name{read_db_tbls}
\alias{read_db_tbls}
\title{Read tables from projects tables.sqlite data base for certain file groups
and save them into a named list}
\usage{
read_db_tbls(db, name_str, name_rmv)
}
\arguments{
\item{db}{Data base object}

\item{name_str}{Text strings which have to part of the table names to
include a table in reading}

\item{name_rmv}{String which should be removed from list element names}
}
\description{
Read tables from projects tables.sqlite data base for certain file groups
and save them into a named list
}
\keyword{internal}
