% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_connectivity.R
\name{reduce_land_connections}
\alias{reduce_land_connections}
\title{Reduce the number of connections between land objects.}
\usage{
reduce_land_connections(data_path, frc_thres)
}
\arguments{
\item{data_path}{Path of the project data folder}

\item{frc_thres}{Fraction threshold to remove connections with a
relative share of connectivity lower than the threshold value.}
}
\value{
The table with the land object connections where connections with
fractions lower than \code{frc_thres} were removed.
}
\description{
Remove connections with connectivity fractions lower than frc_thres.
This is necessary, because keeping all connections between objects
increases the risk of circuit routing (water is routed in circles without
ever ending up in channel or reservoir).
}
