% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_rows_all_na}
\alias{remove_rows_all_na}
\title{Remove rows from table where all elements are NA}
\usage{
remove_rows_all_na(tbl)
}
\arguments{
\item{tbl}{Tibble or data.frame}
}
\description{
Remove rows from table where all elements are NA
}
\keyword{internal}
