% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite_loops.R
\name{resolve_loop_issues}
\alias{resolve_loop_issues}
\title{Resolve infinite loops that occur in the routing between land objects.}
\usage{
resolve_loop_issues(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
The cleaned connections of the land objects.
}
\description{
The routine filters all connections between land objects that occur
in the identified infinite loops, ranks them based on lowest connection
fraction, number of occurrences in all infinite loops and area of the
object that routes (to prefer smaller objects for removing connections).
Connections are removed iteratively until no infinite loop remains.
If a routing fraction is < 1 (object also routes to other objects) a
connection can be removed.
If a fraction is 1 (the object has only this connection) removing this
connection is skipped.
If the routine cannot resolve all infinite loops without removing
connections with a fraction = 1 then an error is triggered and the
issues have to be resolved manually in the land input layer.
}
