% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_soil.R
\name{save_soil_raster}
\alias{save_soil_raster}
\title{Prepare and save the soil raster layer in the project data/raster folder.}
\usage{
save_soil_raster(soil, data_path)
}
\arguments{
\item{soil}{Soil raster layer}

\item{data_path}{Path of the project data folder}
}
\value{
The soil prepared soil raster layer as a .tif file.
}
\description{
Prepare and save the soil raster layer in the project data/raster folder.
}
