% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_connectivity.R
\name{select_point_sub}
\alias{select_point_sub}
\title{Identify the coordinates to where current position points.}
\usage{
select_point_sub(pnt_i, sub)
}
\arguments{
\item{pnt_i}{Numeric value of D8 pointer direction}

\item{sub}{Vector of length 2, row and col index of matrix}
}
\value{
the raster value from the matrix sub based on the pointer value
pnt_i.
}
\description{
Identify the coordinates to where current position points based on the
current matrix coordinates their pointer direction value.
}
\keyword{internal}
