% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{set_proj_crs}
\alias{set_proj_crs}
\title{Write the project CRS into the tables data base}
\usage{
set_proj_crs(bound, data_path)
}
\arguments{
\item{bound}{Basin boundary vector layer}

\item{data_path}{Path of the project data folder}
}
\value{
Writes the table 'project.wkt' into the temporary data base
'/tables.sqlite'.
}
\description{
Write the project CRS into the tables data base
}
