% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hru.R
\name{split_land_layer}
\alias{split_land_layer}
\title{Split the land input layer into the hru and the reservoir layers.}
\usage{
split_land_layer(data_path)
}
\arguments{
\item{data_path}{Path of the project data folder}
}
\value{
Two shape file layers in the folcer vector:
\itemize{
\item hru.shp contains all land polygons
\item res.shp contains all 'watr' polygons which will be reservoirs.
}
}
\description{
Split the land input layer into the hru and the reservoir layers.
}
