% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_par_value}
\alias{translate_par_value}
\title{Translate a label to a respective value.}
\usage{
translate_par_value(x, pot, val)
}
\arguments{
\item{x}{Character vector with the labels to translate}

\item{val}{Numeric vector with the parameter values corresponding to
the labels.}

\item{lbl}{Character vector with the labels corresponding to the
parameter values.}
}
\description{
Translate a label to a respective value.
}
\keyword{internal}
