% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer.R
\name{translate_to_range}
\alias{translate_to_range}
\title{Translate a vector into a vector with the min and max value of that vector.
The max value has a - sign then. This is required to define the id sequences
in the aqu_cha.lin file.}
\usage{
translate_to_range(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Translate a vector into a vector with the min and max value of that vector.
The max value has a - sign then. This is required to define the id sequences
in the aqu_cha.lin file.
}
\keyword{internal}
