% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_source.R
\name{update_ymd_cols}
\alias{update_ymd_cols}
\title{Update the date columns of a point source data table with respect to the
provided time interval.}
\usage{
update_ymd_cols(tbl, int)
}
\arguments{
\item{tbl}{Tibble of user input point source data.}

\item{int}{Time interval, character value which must be one of 'const',
'yr', 'mo', or 'day'}
}
\description{
Update the date columns of a point source data table with respect to the
provided time interval.
}
\keyword{internal}
