% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_shp}
\alias{write_shp}
\title{Write vector layer as shapefile polygon}
\usage{
write_shp(layer, name, path)
}
\arguments{
\item{layer}{Vector layer}

\item{name}{Name of the shapefile layer}

\item{path}{Path to write shape file to.}
}
\description{
Write vector layer as shapefile polygon
}
\keyword{internal}
