% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks_settings.R
\name{check_polygon_topology}
\alias{check_polygon_topology}
\title{Check different polygon properties before writing it into the project
data folder.}
\usage{
check_polygon_topology(
  layer,
  data_path,
  label,
  n_feat = NULL,
  area_fct = 0,
  cvrg_frc = 99.9,
  checks = rep(TRUE, 8)
)
}
\arguments{
\item{layer}{Vector polygon layer}

\item{data_path}{Path of the project data folder}

\item{label}{Layer name to print in error message}

\item{n_feat}{Integer value which defines the exact number of features
the layer must have}

\item{area_fct}{Numeric value to define what amount of the Q25 of area
should be used to calculate threshold for small polygons}

\item{checks}{Boolean vector of length 8 to turn on and off specific
checks}

\item{cvrg_fct}{Numeric value to define the coverage fraction of the
layer with respect to the basin boundary area}
}
\value{
Shape file layers which indicate issues in case of identified issues.
}
\description{
The function intersects the input polygon layer with the basin
boundary layer.
With the intersected layer toplogical checks are performed to check
for MULTIPOLYGON features, small polygons, covered, and overlapping
features.
The function returns the intersected layer if all checks ran
without any identified issues.
Otherwise vector layers are written to the 'vct_path' containing the
features that caused issues in the respective checks.
If at least one check identifies an issue, an error is triggered.
}
